+++ R/sleeper_userleagues.R
@@ -13,17 +13,23 @@
 #'
 #' @export
 
-ff_userleagues.sleeper_conn <- function(conn,user_name = NULL,...){
+ff_userleagues.sleeper_conn <- function(conn, user_name = NULL, ...) {
+  if (is.null(user_name)) {
+    user_id <- conn$user_id
+  }
 
-  if(is.null(user_name)){user_id <- conn$user_id}
+  if (!is.null(user_name)) {
+    user_id <- .sleeper_userid(user_name)
+  }
 
-  if(!is.null(user_name)){user_id <- .sleeper_userid(user_name)}
-
-  df_leagues <- sleeper_getendpoint("user",user_id,"leagues/nfl",conn$season) %>%
+  df_leagues <- sleeper_getendpoint("user", user_id, "leagues/nfl", conn$season) %>%
     purrr::pluck("content") %>%
-    purrr::map_dfr(`[`,c("name","league_id")) %>%
-    dplyr::mutate(franchise_name = purrr::map_chr(.data$league_id,.sleeper_userteams,user_id),
-                  franchise_id = user_id)
+    purrr::map_dfr(`[`, c("name", "league_id")) %>%
+    dplyr::rename(league_name = .data$name) %>%
+    dplyr::mutate(
+      franchise_name = purrr::map_chr(.data$league_id, .sleeper_userteams, user_id),
+      franchise_id = user_id
+    )
 
   return(df_leagues)
 }
@@ -31,14 +37,13 @@ ff_userleagues.sleeper_conn <- function(conn,user_name = NULL,...){
 #' Get User Teams
 #' @noRd
 
-.sleeper_userteams <- function(league_id,user_id){
-
-  df_teams <- sleeper_getendpoint("league",league_id,"users") %>%
+.sleeper_userteams <- function(league_id, user_id) {
+  df_teams <- sleeper_getendpoint("league", league_id, "users") %>%
     purrr::pluck("content") %>%
     tibble::tibble() %>%
-    tidyr::hoist(1,"franchise_id"="user_id","display_name","metadata") %>%
-    tidyr::hoist("metadata","franchise_name" = "team_name") %>%
-    dplyr::mutate("franchise_name"=dplyr::coalesce(.data$franchise_name,.data$display_name)) %>%
+    tidyr::hoist(1, "franchise_id" = "user_id", "display_name", "metadata") %>%
+    tidyr::hoist("metadata", "franchise_name" = "team_name") %>%
+    dplyr::mutate("franchise_name" = dplyr::coalesce(.data$franchise_name, .data$display_name)) %>%
     dplyr::filter(.data$franchise_id == user_id) %>%
     dplyr::pull("franchise_name")