+++ tests/testthat/helper.R
@@ -1,3 +1,4 @@
 suppressPackageStartupMessages({
   library(httptest)
+  library(checkmate)
 })