+++ tests/testthat/test-ff_draftpicks.R
@@ -6,10 +6,17 @@ with_mock_api({
     sfb <- mfl_connect(2020, 65443)
     sfb_picks <- ff_draftpicks(sfb)
 
-    expect_s3_class(ssb_picks, "tbl_df")
-    expect_gt(nrow(ssb_picks), 1)
+    expect_tibble(ssb_picks,min.rows = 1)
+    expect_tibble(sfb_picks, nrows = 0)
 
-    expect_s3_class(sfb_picks, "tbl_df")
-    expect_equal(nrow(sfb_picks), 0)
+    jml_conn <- ff_connect("sleeper", 522458773317046272, season = 2020)
+    jml_picks <- ff_draftpicks(jml_conn)
+
+    dlp <- sleeper_connect(2020, 521379020332068864)
+    dlp_picks <- ff_draftpicks(dlp)
+
+    expect_tibble(jml_picks,min.rows = 144)
+    expect_tibble(dlp_picks, min.rows = 144)
+
   })
 })