+++ tests/testthat/test-ff_rosters.R
@@ -1,11 +1,11 @@
 with_mock_api({
   test_that("ff_rosters returns a tibble", {
-
-    ssb <- mfl_connect(2020,54040)
+    ssb <- mfl_connect(2020, 54040)
     ssb_rosters <- ff_rosters(ssb)
+    expect_tibble(ssb_rosters, min.rows = 300)
 
-    expect_s3_class(ssb_rosters,"tbl_df")
-    expect_gt(nrow(ssb_rosters),1)
-
+    jml_conn <- sleeper_connect(league_id = 522458773317046272, season = 2020)
+    jml_rosters <- ff_rosters(jml_conn)
+    expect_tibble(jml_rosters, min.rows = 200)
   })
 })