+++ tests/testthat/test-ff_schedule.R
@@ -1,20 +1,20 @@
 with_mock_api({
-  test_that("ff_schedule returns a tibble",{
-
-    dlf <- mfl_connect(2019,37920)
+  test_that("ff_schedule returns a tibble", {
+    dlf <- mfl_connect(2019, 37920)
     dlf_schedule <- ff_schedule(dlf)
 
-    expect_s3_class(dlf_schedule,"tbl_df")
-    expect_gt(nrow(dlf_schedule),1)
-
-    ssb <- mfl_connect(2020,54040)
+    ssb <- mfl_connect(2020, 54040)
     ssb_schedule <- ff_schedule(ssb)
 
-    fog <- mfl_connect(2019,12608)
+    fog <- mfl_connect(2019, 12608)
     fog_schedule <- ff_schedule(fog)
 
-    # expect_s3_class(fog_schedule,"tbl_df")
-    expect_null(fog_schedule)
+    jml_conn <- sleeper_connect(league_id = 522458773317046272, season = 2020)
+    jml_schedule <- ff_schedule(jml_conn)
 
+    expect_tibble(ssb_schedule, min.rows = 100)
+    expect_tibble(dlf_schedule, min.rows = 100)
+    expect_tibble(jml_schedule, min.rows = 100)
+    expect_null(fog_schedule)
   })
 })